+++ CHANGELOG.md
@@ -7,7 +7,7 @@
 
 ## Unreleased: mitmproxy next
 
-- Add `--set save_stream_compression` option to compress stream files on the fly with zstandard.
+- Add `--set save_stream_compress` option to compress stream files on the fly with zstandard.
   Reading compressed flow files is auto-detected via magic bytes.
   ([#8287](https://github.com/mitmproxy/mitmproxy/pull/8287), @joelverhagen)
 - Bracket IPv6 target literals in the `CONNECT` request and `Host` header sent